Always generate files in closure_grpc_web_library()
This adds a `ctx.actions.write` to each of the automatically generated source targets
that have been declared.  This prevents an issue if the source targets do not yield a
`.grpc.js` generated file because they do not contain an `rpc` definition (e.g. proto
files that only contain a `message`).

Resolves: #1076